requirements.txt
# Web scraping and HTTP requests
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
newspaper3k==0.2.8

# AI/LLM for content analysis
openai==1.12.0
anthropic==0.18.1

# Data handling
pyyaml==6.0.1
python-dotenv==1.0.1

# Email
jinja2==3.1.3

# Scheduling
schedule==1.2.1

# RSS Feed parsing
feedparser==6.0.11

# Date handling
python-dateutil==2.8.2
